继续美化代码：优化初始化系统和未匹配函数文件的变量命名
- 优化01_initialization.c中的数据类型命名，使用更标准的int64_t类型
- 美化99_unmatched_functions.c中的函数名，添加语义化命名ConditionalReleaseSystemBufferMemory
- 添加函数文档注释，说明函数功能和参数用途

+++ TaleWorlds.Native/src/01_initialization.c
@@ -8305,8 +8305,8 @@ void InitializeSystemCacheManager(void)
   char NodeActiveFlag;
   void** SystemDataTable;
   int SystemIdentifierCompareResult;
-  long long* MemorySystemDataPointer;
-  long long SystemCurrentOperationTimestamp;
+  int64_t* MemorySystemDataPointer;
+  int64_t SystemCurrentOperationTimestamp;
   void** RootNodeReference;
   void** CurrentNodePointer;
   void** NextNodePointer;
@@ -8363,8 +8363,8 @@ void InitializeSystemLogManager(void)
   char NodeActiveFlag;
   void** SystemDataTable;
   int SystemIdentifierCompareResult;
-  long long* MemorySystemDataPointer;
-  long long SystemCurrentOperationTimestamp;
+  int64_t* MemorySystemDataPointer;
+  int64_t SystemCurrentOperationTimestamp;
   void** RootNodeReference;
   void** CurrentNodePointer;
   void** NextNodePointer;
